refactor(simulation): split shared formatting from numeric helpers
